requirements.txt
pyserial
persistqueue
influxdb-client
requests
RPi.GPIO
spidev
python-dotenv
